Don't break if the directory doesn't exist. PLUG-606
git-svn-id: https://studio.atlassian.com/svn/PLUG/trunk@62201 2c54a935-e501-0410-bc05-97a93f6bca70

+++ atlassian-plugins-core/src/main/java/com/atlassian/plugin/util/zip/AbstractUnzipper.java
@@ -87,7 +87,10 @@ public abstract class AbstractUnzipper implements Unzipper
         if (!targetDirContents.equals(zipContentsAndLastModified))
         {
             // Note: clean, not delete, as destdir may be a symlink (PLUG-606).
-            FileUtils.cleanDirectory(destDir);
+            if (destDir.exists())
+            {
+                FileUtils.cleanDirectory(destDir);
+            }
             unzip();
         }
         else
